[lgdo] refactor subpackage
* move LH5 code into own subpackage lgdo/lh5 and rename modules
* create lgdo.utils and lgdo.lh5.utils
* update all import statements in pygama
