+++ src/pygama/__init__.py
@@ -3,6 +3,7 @@ Pygama: decoding and processing digitizer data.
 Check out the `online documentation <https://pygama.readthedocs.io>`_
 """
 
-from pygama._version import version as __version__
+from ._version import version as __version__
+from .lgdo import lh5
 
-__all__ = ["__version__"]
+__all__ = ["__version__", "lh5"]
