+++ src/pygama/hit/build_hit.py
@@ -10,7 +10,7 @@ from collections import OrderedDict
 
 import numpy as np
 
-from pygama.lgdo import LH5Iterator, LH5Store, ls
+from pygama.lgdo.lh5 import LH5Iterator, LH5Store, ls
 
 log = logging.getLogger(__name__)
