+++ src/pygama/lgdo/__init__.py
@@ -29,21 +29,14 @@ browsed easily in python like any `HDF5 <https://www.hdfgroup.org>`_ file using
 `h5py <https://www.h5py.org>`_.
 """
 
-from pygama.lgdo.array import Array
-from pygama.lgdo.arrayofequalsizedarrays import ArrayOfEqualSizedArrays
-from pygama.lgdo.fixedsizearray import FixedSizeArray
-from pygama.lgdo.lh5_store import LH5Iterator, LH5Store, load_dfs, load_nda, ls, show
-from pygama.lgdo.scalar import Scalar
-from pygama.lgdo.struct import Struct
-from pygama.lgdo.table import Table
-from pygama.lgdo.vectorofvectors import (
-    VectorOfVectors,
-    build_cl,
-    explode,
-    explode_arrays,
-    explode_cl,
-)
-from pygama.lgdo.waveform_table import WaveformTable
+from .array import Array
+from .arrayofequalsizedarrays import ArrayOfEqualSizedArrays
+from .fixedsizearray import FixedSizeArray
+from .scalar import Scalar
+from .struct import Struct
+from .table import Table
+from .vectorofvectors import VectorOfVectors
+from .waveform_table import WaveformTable
 
 __all__ = [
     "Array",
@@ -54,16 +47,6 @@ __all__ = [
     "Table",
     "VectorOfVectors",
     "WaveformTable",
-    "LH5Iterator",
-    "LH5Store",
-    "load_dfs",
-    "load_nda",
-    "ls",
-    "show",
-    "build_cl",
-    "explode",
-    "explode_arrays",
-    "explode_cl",
 ]
 
 import numpy as np
