+++ src/pygama/lgdo/utils.py
@@ -3,9 +3,7 @@ Implements utilities for LEGEND Data Objects.
 """
 from __future__ import annotations
 
-import glob
 import logging
-import os
 
 import numpy as np
 
@@ -88,32 +86,3 @@ def parse_datatype(datatype: str) -> tuple[str, tuple[int, ...], str | list[str]
         return datatype, tuple(dims), element_description
     else:
         return datatype, None, element_description.split(",")
-
-
-def expand_path(path: str, list: bool = False) -> str | list:
-    """Expand environment variables and wildcards to return absolute path
-
-    Parameters
-    ----------
-    path
-        name of path, which may include environment variables and wildcards
-    list
-        if True, return a list. If False, return a string; if False and a
-        unique file is not found, raise an Exception
-
-    Returns
-    -------
-    path or list of paths
-        Unique absolute path, or list of all absolute paths
-    """
-
-    paths = glob.glob(os.path.expanduser(os.path.expandvars(path)))
-    if not list:
-        if len(paths) == 0:
-            raise FileNotFoundError(f"could not find path matching {path}")
-        elif len(paths) > 1:
-            raise FileNotFoundError(f"found multiple paths matching {path}")
-        else:
-            return paths[0]
-    else:
-        return paths
