+++ src/pygama/pargen/AoE_cal.py
@@ -21,11 +21,11 @@ from matplotlib.backends.backend_pdf import PdfPages
 from matplotlib.colors import LogNorm
 from scipy.stats import chi2
 
-import pygama.lgdo.lh5_store as lh5
 import pygama.math.histogram as pgh
 import pygama.math.peak_fitting as pgf
 import pygama.pargen.ecal_th as thc
 import pygama.pargen.energy_cal as pgc
+from pygama.lgdo import lh5
 
 log = logging.getLogger(__name__)
