+++ src/pygama/pargen/cuts.py
@@ -13,10 +13,10 @@ import numpy as np
 import pandas as pd
 from scipy import stats
 
-import pygama.lgdo.lh5_store as lh5
 import pygama.math.histogram as pgh
 import pygama.math.peak_fitting as pgf
 import pygama.pargen.energy_cal as pgc
+from pygama.lgdo import lh5
 
 log = logging.getLogger(__name__)
