+++ tests/hit/test_build_hit.py
@@ -4,9 +4,8 @@ from pathlib import Path
 import numpy as np
 import pytest
 
-import pygama.lgdo.lh5_store as store
 from pygama.hit import build_hit
-from pygama.lgdo import LH5Store, ls
+from pygama.lgdo import lh5
 
 config_dir = Path(__file__).parent / "configs"
 
@@ -22,7 +21,7 @@ def test_basics(dsp_test_file):
     )
 
     assert os.path.exists(outfile)
-    assert ls(outfile, "/geds/") == ["geds/hit"]
+    assert lh5.ls(outfile, "/geds/") == ["geds/hit"]
 
 
 def test_illegal_arguments(dsp_test_file):
@@ -57,7 +56,7 @@ def test_lh5_table_configs(dsp_test_file):
     )
 
     assert os.path.exists(outfile)
-    assert ls(outfile, "/geds/") == ["geds/hit"]
+    assert lh5.ls(outfile, "/geds/") == ["geds/hit"]
 
     lh5_tables_config = {
         "/geds/dsp": {
@@ -80,7 +79,7 @@ def test_lh5_table_configs(dsp_test_file):
     )
 
     assert os.path.exists(outfile)
-    assert ls(outfile, "/geds/") == ["geds/hit"]
+    assert lh5.ls(outfile, "/geds/") == ["geds/hit"]
 
 
 def test_outputs_specification(dsp_test_file):
@@ -93,7 +92,7 @@ def test_outputs_specification(dsp_test_file):
         wo_mode="overwrite",
     )
 
-    store = LH5Store()
+    store = lh5.LH5Store()
     obj, _ = store.read_object("/geds/hit", outfile)
     assert list(obj.keys()) == ["calE", "AoE", "A_max"]
 
@@ -106,9 +105,9 @@ def test_build_hit_spms_basic(dsp_test_file_spm):
         hit_config=f"{config_dir}/spms-hit-config.json",
         wo_mode="overwrite_file",
     )
-    assert ls(out_file) == ["ch0", "ch1", "ch2"]
-    assert ls(out_file, "ch0/") == ["ch0/hit"]
-    assert ls(out_file, "ch0/hit/") == [
+    assert lh5.ls(out_file) == ["ch0", "ch1", "ch2"]
+    assert lh5.ls(out_file, "ch0/") == ["ch0/hit"]
+    assert lh5.ls(out_file, "ch0/hit/") == [
         "ch0/hit/energy_in_pe",
         "ch0/hit/quality_cut",
         "ch0/hit/trigger_pos",
@@ -124,9 +123,9 @@ def test_build_hit_spms_multiconfig(dsp_test_file_spm):
         lh5_tables_config=f"{config_dir}/spms-hit-multi-config.json",
         wo_mode="overwrite",
     )
-    assert ls(out_file) == ["ch0", "ch1", "ch2"]
-    assert ls(out_file, "ch0/") == ["ch0/hit"]
-    assert ls(out_file, "ch0/hit/") == [
+    assert lh5.ls(out_file) == ["ch0", "ch1", "ch2"]
+    assert lh5.ls(out_file, "ch0/") == ["ch0/hit"]
+    assert lh5.ls(out_file, "ch0/hit/") == [
         "ch0/hit/energy_in_pe",
         "ch0/hit/quality_cut",
         "ch0/hit/trigger_pos",
@@ -142,13 +141,13 @@ def test_build_hit_spms_calc(dsp_test_file_spm):
         wo_mode="overwrite_file",
         lh5_tables_config=f"{config_dir}/spms-hit-a-config.json",
     )
-    assert ls(out_file) == ["ch0", "ch1", "ch2"]
-    assert ls(out_file, "ch0/") == ["ch0/hit"]
-    assert ls(out_file, "ch0/hit/") == ["ch0/hit/energy_in_pe"]
+    assert lh5.ls(out_file) == ["ch0", "ch1", "ch2"]
+    assert lh5.ls(out_file, "ch0/") == ["ch0/hit"]
+    assert lh5.ls(out_file, "ch0/hit/") == ["ch0/hit/energy_in_pe"]
 
-    df0 = store.load_nda(out_file, ["energy_in_pe"], "ch0/hit/")
-    df1 = store.load_nda(out_file, ["energy_in_pe"], "ch1/hit/")
-    df2 = store.load_nda(out_file, ["energy_in_pe"], "ch2/hit/")
+    df0 = lh5.load_nda(out_file, ["energy_in_pe"], "ch0/hit/")
+    df1 = lh5.load_nda(out_file, ["energy_in_pe"], "ch1/hit/")
+    df2 = lh5.load_nda(out_file, ["energy_in_pe"], "ch2/hit/")
 
     assert len(df0["energy_in_pe"]) == 5
     assert len(df1["energy_in_pe"]) == 5
