+++ tests/raw/test_build_raw.py
@@ -3,7 +3,7 @@ from pathlib import Path
 
 import pytest
 
-from pygama.lgdo.lh5_store import LH5Store, ls
+from pygama.lgdo.lh5 import LH5Store, ls
 from pygama.raw import build_raw
 
 config_dir = Path(__file__).parent / "configs"
